requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
autopep8==1.4.4
Babel==2.9.0
black==20.8b1
boilerplate==1.2.4
certifi==2020.12.5
cffi==1.14.6
chardet==4.0.0
click==7.1.2
colorama==0.4.3
cryptography==3.4.7
defusedxml==0.7.1
Deprecated==1.2.10
diff-match-patch==20200713
distlib==0.3.1
dj-database-url==0.5.0
Django==2.2.13
django-allauth==0.45.0
django-appconf==1.0.4
django-crispy-forms==1.12.0
django-extra-views==0.13.0
django-filter==2.4.0
django-haystack==3.0
django-heroku==0.3.1
django-import-export==2.6.1
django-oscar==3.0
django-phone-verify==2.0.0
django-phonenumber-field==3.0.1
django-tables2==2.3.4
django-treebeard==4.4
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
et-xmlfile==1.1.0
factory-boy==2.12.0
Faker==6.0.0
filelock==3.0.12
Flask==1.1.2
gunicorn==20.1.0
hupper==1.10.2
idna==2.10
isort==5.5.2
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
Markdown==3.3.4
MarkupPy==1.14
MarkupSafe==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
nexmo==2.5.2
oauthlib==3.1.1
odfpy==1.4.1
openpyxl==3.0.9
PasteDeploy==2.1.0
pathspec==0.8.1
pep8==1.7.1
phonenumbers==8.12.15
phonenumberslite==8.12.35
Pillow==8.1.0
plaster==1.0
plaster-pastedeploy==0.7
psycopg2==2.8.6
purl==1.5
pycodestyle==2.5.0
pycparser==2.20
PyJWT==2.0.0
pylint==2.6.0
pyramid==1.10.4
python-dateutil==2.8.1
python-dotenv==0.15.0
python3-openid==3.2.0
pytz==2018.5
pywin32==302
PyYAML==6.0
redis==3.5.3
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
Rest_Framework_Auth_Toolkit==0.10
rjsmin==1.1.0
six==1.15.0
sqlparse==0.2.4
tablib==3.0.0
text-unidecode==1.3
toml==0.10.1
translationstring==1.4
twilio==6.50.1
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.2
venusian==3.0.0
virtualenv==20.0.31
virtualenvwrapper-win==1.2.6
WebOb==1.8.6
Werkzeug==1.0.1
whitenoise==5.3.0
wrapt==1.12.1
xlrd==2.0.1
xlwt==1.3.0
zope.deprecation==4.4.0
zope.interface==5.1.0
